+++ hypothesis-python/RELEASE.rst
@@ -1,0 +1,4 @@
+RELEASE_TYPE: patch
+
+This patch moves some internal code, so that future work can avoid
+creating import cycles.  There is no user-visible change.
